+++ src/scripts/components/Activity.js
